Test DB configured. Packaging updated.
