--- src/main/java/com/workfront/internship/dao/PostDAO.java
@@ -1,38 +1,0 @@
-package com.workfront.internship.dao;
-
-import com.workfront.internship.dataModel.Post;
-
-import java.util.List;
-
-/**
- * Created by nane on 5/27/17.
- */
-public interface PostDAO {
-
-    long add(Post post);
-
-    /**
-     * Gets all Posts from database.
-     *
-     * @return
-     */
-    List<Post> getAll();
-
-    List<Post> getByUserId(long userId);
-
-    List<Post> getByTitle(String title);
-
-    Post getById(long id);
-
-    List<Post> getAnswersByPostId(long postId);
-
-    Post getBestAnswer(long postId);
-
-    boolean setBestAnswer(long postId, long answerId);
-
-    boolean update(Post post);
-
-    int delete(long id);
-
-
-}
