--- src/main/java/com/workfront/internship/dao/UserDAO.java
@@ -1,72 +1,0 @@
-package com.workfront.internship.dao;
-
-import com.workfront.internship.dataModel.AppArea;
-import com.workfront.internship.dataModel.User;
-
-import java.util.List;
-
-
-public interface UserDAO {
-
-    //CREATE
-
-    /**
-     * Adds user to the database
-     * @param user
-     * @return
-     */
-    long add(User user);
-
-    //READ
-
-    /**
-     * Gets the list of the users by the given name
-     * @param name
-     * @return
-     */
-    List<User> getByName(String name);
-
-    /**
-     * Gets the user by the given id
-     * @param id
-     * @return
-     */
-    User getById(long id);
-
-    /**
-     * Gets the list of AppAreas which the user with the given id is subscribed to
-     * @param id
-     * @return
-     */
-    List<AppArea> getAppAreasById(long id);
-
-    //UPDATE
-
-    /**
-     * Subscribes the given user to the given appArea
-     * @param userId
-     * @param appAreaId
-     */
-    void subscribeToArea(long userId, long appAreaId);
-
-    /**
-     * Removes the subscription of the given user to the given appArea
-     * @param userId
-     * @param appAreaId
-     */
-    void unsubscribeToArea(long userId, long appAreaId);
-
-    //DELETE
-
-    /**
-     * Deletes the user with the given id
-     * @param id
-     */
-    void deleteById(long id);
-
-    /**
-     * Deletes all the existing users
-     */
-    void deleteAll();
-
-}
