--- src/main/java/com/workfront/internship/dataModel/Comment.java
@@ -1,62 +1,0 @@
-package com.workfront.internship.dataModel;
-
-import java.sql.Date;
-
-
-public class Comment {
-    private long id;
-    private User user;
-    private Post post;
-    private String content;
-    private Date dateTime;
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public User getUser() {
-        return user;
-    }
-
-    public void setUser(User user) {
-        this.user = user;
-    }
-
-    public Post getPost() {
-        return post;
-    }
-
-    public void setPost(Post post) {
-        this.post = post;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-
-    public Date getDateTime() {
-        return dateTime;
-    }
-
-    public void setDateTime(Date dateTime) {
-        this.dateTime = dateTime;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        return this == o || (o instanceof Comment) && id == ((Comment) o).id;
-    }
-
-    @Override
-    public int hashCode() {
-        return (int) (id ^ (id >>> 32));
-    }
-}
