--- src/main/java/com/workfront/internship/exceptions/dao/NotExistingAppAreaException.java
@@ -1,16 +1,0 @@
-package com.workfront.internship.exceptions.dao;
-
-/**
- * Created by Vahag on 6/2/2017.
- */
-public class NotExistingAppAreaException extends RuntimeException {
-
-    // Parameterless Constructor
-    public NotExistingAppAreaException() {}
-
-    // Constructor that accepts a message
-    public NotExistingAppAreaException(String message)
-    {
-        super(message);
-    }
-}
