+++ src/main/resources/dbConfig.properties
@@ -1,3 +1,4 @@
-URL = jdbc:mysql://localhost:3306/work_flow
+DB_URL = jdbc:mysql://localhost:3306/work_flow
+DB_TEST_URL = jdbc:mysql://localhost:3306/work_flow_test
 USERNAME = root
 PASSWORD = vahag1234
